+++ packages/components/src/components/Initials/stories/Default.stories.tsx
@@ -1,6 +1,8 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import Initials from "../Initials";
 import React from "react";
+import { Flex } from "@/components/Flex";
+import { Avatar } from "@/components/Avatar";
 
 const meta: Meta<typeof Initials> = {
   title: "Content/Initials",
@@ -18,4 +20,14 @@ export const Default: Story = {};
 
 export const OneLetter: Story = {
   render: (props) => <Initials {...props}>Max </Initials>,
+};
+
+export const Emoji: Story = {
+  render: (props) => (
+    <Flex direction="column">
+      <Avatar>
+        <Initials {...props}>😄</Initials>
+      </Avatar>
+    </Flex>
+  ),
 };