+++ packages/components/src/index/default.ts
@@ -5,3 +5,4 @@ export * from "@/lib/hooks/public";
 export * from "@/lib/propsContext/index";
 export * from "@/lib/remote/eventHandlerContext";
 export { useControlledRemoteValueProps } from "@/lib/remote/useControlledRemoteValueProps";
+export { flags } from "@/flags";